fix(oms): implement 64-bit bit-packing for client order id
- Switch Decompose parameter from uint to ulong to prevent truncation.
- Add OMS_MASK (bit 63) to distinguish live orders from historical data.
- Partition 64-bit space into: [1-bit Mask][31-bit Account][32-bit Seq].
- Ensure Decompose correctly strips the MSB before retrieving accountId.

+++ OMS/src/Trading.Oms.Domain/Interface/IOrderIdComposer.cs
@@ -4,5 +4,5 @@ public interface IOrderIdComposer
 {
     ulong Compose(uint accountId, uint sequence);
 
-    (uint accountId, uint sequence) Decompose(uint orderId);
+    (uint accountId, uint sequence) Decompose(ulong orderId);
 }
